+++ CotEditor/Sources/EditorTextView.swift
@@ -558,7 +558,7 @@ final class EditorTextView: NSTextView, Themable {
         }
         
         // add "Inspect Character" menu item if single character is selected
-        if (self.string as NSString).substring(with: self.selectedRange).numberOfComposedCharacters == 1 {
+        if (self.string as NSString).substring(with: self.selectedRange).count == 1 {
             menu.insertItem(withTitle: NSLocalizedString("Inspect Character", comment: ""),
                             action: #selector(showSelectionInfo(_:)),
                             keyEquivalent: "",
@@ -902,8 +902,7 @@ final class EditorTextView: NSTextView, Themable {
             return self.selectedRange.length > 0
             
         case #selector(showSelectionInfo):
-            let selection = (self.string as NSString).substring(with: self.selectedRange)
-            return selection.numberOfComposedCharacters == 1
+            return (self.string as NSString).substring(with: self.selectedRange).count == 1
             
         case #selector(toggleComment):
             if let menuItem = item as? NSMenuItem {